+++ java-desktop-book/code/src/com/googlecode/projecttemplate/pos/service/PersonService.java
@@ -17,7 +17,7 @@ public interface PersonService {
     void save(Person person);
     void delete(Person person);
     Long count();
-    Person getPerson(Integer id);
+    Person getPerson(Long id);
     List<Person> getPersons();
     List<Person> getPersons(int start, int num);
